Creation of the HelpCameraSettingsScene Update of GameManager and index to include the new scene Change title of helpCameraSettings Suppression function reset
